Ignore errors for tag repository
While it is great to have this repo around, in case
of EOL releases it can be inaccessible. In case it is
not accesible the only thing happens is we can have
a bit more older packages around, what is not such
a mess, because Fedora is updated a lot.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>