.envrc: unset VIRTUAL_ENV
If I have a random venv enabled when I enter direnv it
blows up in my face and the random venv is kept.
This fixes the problem for me and the expected poetry
venv works.

I tried to use `deactivate` but it is not around inside
`.envrc` not sure why ...

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>