jobs: Add rsync options to archive()
Drop the redundant `-r` rsync option (it's already part of `-a`).

Drop the `-P` option from the RHEL ranch. This isn't useful for writing
into log files, and was an unnecessary inconsistency with the public
jobs.

Most calls to archive() are to individual files, except for the task
logs. We are going to change that in the next step and need more
flexibility: So change archive() to *not* copy files recursively by
default, and introduce a third argument to archive() for passing
additional options to rsync. Use that in the "sync task logs" loop to
explicitly pass `--recursive` to keep the current behaviour for now.

Spell out the rsync options to make them easier to understand.

Robustify the quoting.