jobs: Skip archiving TMT dist-git checkout
Start picking out only selected files/directories for archiving, as
always archiving the full work space is a giant waste of space and leads
to running out of disk space or inodes [1].

As a first step, only archive all top-level files, the guest setup logs,
and the TMT output dir. In particular, this skips "workdir-*", which is
test_schedule_tmt's dist-git checkout. This can easily be retrieved
locally with just a git clone, there is nothing at all interesting about
it.

[1] https://issues.redhat.com/browse/TFT-1735