job: Fix redhat ranch rsyncing of TMT results
Commit d9b8db55828f accidentally changed the rsync destination variables
in the redhat TMT jobs to the same ones as on the public ranch
(copy pasta accident). As a quick fix, just put back the previous rsync
destination.

This is an on-going trap door. Cleaning this up properly will require
some more thought, let's not do that in a hurry.

https://issues.redhat.com/browse/TFT-1749