Use google DNS in CI
Currently, gitlab-runner is deployed on the same network as
staging/production.

Because of this, the kube API hostname is resolved as non-public
IP address, which is not reachable of CI, as they are in different
security groups.

We plan to move gitlab-runner deployment away anyway, and we will host
it on a separate cluster rather, for security resons.

So temporarilty, this seems to be the easiest way to fix
the problem we are hitting.

Tested here:

https://gitlab.com/testing-farm/infrastructure/-/jobs/3981093239

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>