Add terminate-artemis-guests-dev Makefile target
This removes all instances created by Artemis from the dev environment.

Intended to be used in Gitlab CI to cleanup all instances before
Terraform starts to remove the security group. This causes a race
condition if Artemis is not quick enough. We are in this trouble
as those instances are managed by Artemis, not Terraform.

Should resolve problems like:

```
│ Error: deleting Security Group (sg-084fc835b8f5975e7): DependencyViolation: resource sg-084fc835b8f5975e7 has a dependent object
│ 	status code: 400, request id: fc9fc7ab-c642-4fb6-848d-2a26ff1bb8ac
```

When cleaning up dev environment.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>