Add development eks cluster via terragrunt
First patch to move to terragrunt based infrastructure.

This creates cluster, and the cluster is a separate module in the dev
environment.

Have the code in a new `terragrunt` directory, we will drop terraform
parts once the migration ready.

Same goes for the Makfile, add new make teragets to create the
development environment, but keep the old one so folks can destroy
their old development environment.

Change the `test/dev` gitlab job to call the new development
environment.

Drop `test/guest-setup` which is broken anyway, we will bring it later.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>