Fix wait_artemis_available.sh script
This script was before available only for `dev` environment.
Make it more generic and use it also for staging environment.

This script is useful to verify localhost can connect to Artemis
and also to wait until the added hostname is resolvable by DNS.

It will be used as a dependency for test execution, to make sure
connection to Artemis works well.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>