Add artemis-ci deployment in staging
After thinking about this more, seems the most straight forward thing
will be adding a separate deployment of Artemis in the staging
environment for CI testing.

This instance will inherit the configration from the "main" Artemis staging instance but:

* will be provisioned using the local backend
* will be deployed in a generated namespace

This instance Gitlab CI will deploy from the configuration changes, so
we can run integration tests against during a compose update.

Small fixes were needed in the module so that the security group is
unique to the artemis deployment.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>