Run all worker tests against Artemis CI instance deployed on staging
So with the previous patches we are ready for this:

* a fresh Artemis instance is deployed from the MR
* worker tests are run against this instance
* the instance is tierd down

Previously this testing timed out, now it should finish under 30 minutes.
We will optimize this more in the future.

There was a fix needed in `pytest-gluetool` to make the tests stable:

* https://gitlab.com/testing-farm/pytest-gluetool/-/merge_requests/34

And we had to fix the permission for the Artemis private SSH key also,
it was incorrect.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>