citool-config: consider corrupted tmt metadata as user error
After few months of observing these errors, I believe it safe
to say this is a user error the user should care about:

Examples from current error page:
http://artifacts.osci.redhat.com/testing-farm/c6314f6f-5e62-4b1a-95a8-7aafaccae14c/
http://artifacts.osci.redhat.com/testing-farm/d7c10d87-b16a-4b13-b5b8-86972951fc5d/
https://artifacts.dev.testing-farm.io/4c8dc5b4-3ecb-4f9e-a623-b86197e292cf/
https://artifacts.dev.testing-farm.io/19bcd4f2-871c-44bd-9fbb-6f2798262e5c/
http://artifacts.osci.redhat.com/testing-farm/c161f840-ce61-438c-9543-19634cbff1ba/

On current page there are 309 errors out of ~1100 for the last 14 days.

This should significantly lower the error rate. I do believe the error
message is clear enough for user to investigate.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>