Prepare for local development environment
1. Remove poetry activation in direnv

The poetry activation will be causing trouble with local development,
where most of the tools are installed via poetry.
Let's not activate it automatically, just add the virtual
env to the PATH so the installed tools are available.

2. Make the `.vault_pass` optional

For local development environment it is not needed. Rather just
a warning for mataintainers, to add the secret file if they want
to use the infrastructure repository as Testing Farm maintainers.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>