workers: move to Fedora 39 cloud base image_id
So we want to move to newer Fedora as F36 is EOL.

Unfortunately F39 is using Python 3.12, which breaks `get_url` module.
The fix went into Ansible 2.16 but that requires Python 3.10+, and we
agreed on team level to stick to Python3.9 now.

So to resolve the issue, move to `curl` rather, similar fix went into
BaseOS CI some time ago.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>