Factor out getting localhost IP to separate module
We used transitive dependencies in the current setup:

```
server/security-group -> artemis -> eks
```

Where `artemis` module provided the localhost IP.

We are hitting these bugs:

https://github.com/gruntwork-io/terragrunt/issues/3028 (version 0.57.2
and higher).

https://github.com/gruntwork-io/terragrunt/pull/2957 (version 0.55.0 and
lower).

By moving it to a seaprate module we remove this transitive dependency
and do not hit the bugs anymore:

```
server/security-group -> localhost

artemis -> localhost
        |
        -> eks
```

With that update to latest terragrunt `0.57.2`, so we do not hit
additional bugs.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>