Support Testing Farm github runners
With the RHTAS (SecureSign) team we have a new use-case where they would
like to use Testing Farm solely as a infrastructure provider to their
Ansible Molecule tests. They do not want to add there `tmt` wrapping
around molecule to mitigate confusing the upstream folks who are used to
contributing directly to Ansible Molecule.

For this use case the team will be using our reservations which provide
the machine and then pass it over to molecule which will handle the test
execution.

For this to work we need a custom repository specific github runner for them,
which behaves as a testing farm worker and has access to the test
guests. We cannot make this work on generally available GitHub action
runners, because there is like thousands of them, and our security group
would not be able to cope with that list of rules due to limitations.

Also note that we are adjusting the guests security group to allow
all communication from workers to guests, this is needed so the github
runners can freely communicate with the guests without any restrictions.

See README.adoc for documentation for more information.

Resolves TFT-2722

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>