worker-redhat: enable both tag repositories
We are trying an approach where we add both available
tag repositories to the systems (mainline and z-stream)
to mitigate issues with choosing the wrong one.

This simple approach should work well, as we really need
this repository only to be present to provide some additional
dependencies specified by the version requirements.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>