citool-container: mount configuration as overlay
We are distributing `git-credentials` with citool-config with
a default set of credentials use to authenticate against
`gitlab.com `to have access to `gitlab.com/redhat` namespace.

We also support the user to pass credentials via git url
if he wants to access some private repositories.

Because we use the `store` git credentials helper, any
git secret passed via the URL will be automatically saved
and not needed to be used next time when cloning that git
repository. This can be handy that user will be able to discover
additional tests with the same credentials, without the need
to pass the environment variables for tmt.

The problem with the above is that this in current setup
changes the credentials on the worker, as the configuration
is shared between all worker containers.

To mitigate this problem, use an overaly mount, so the changes
to the git credentials file are done only in the running container
and the worker configuration stays intact.

For more info see:

```
man gitcredentials
man git-credential-store
```

Resolves TFT-3927

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>